separate module for executeScript
